Move intents into an 'intent' folder
This commit also moves the 'intents index' logic from the lambda function index into the intents/index.js file.
