+++ oscarbot/intents/countIssues.js
@@ -1,8 +1,8 @@
-const config = require('./config');
-const dialogActions = require('./utils/dialogActions');
-const login = require('./utils/github/login');
-const query = require('./utils/github/query');
-const i18n = require('./i18n');
+const config = require('../config');
+const dialogActions = require('../utils/dialogActions');
+const login = require('../utils/github/login');
+const query = require('../utils/github/query');
+const i18n = require('../i18n');
 
 function handler(event, context, callback) {
   const repository = event.sessionAttributes.Repository;
@@ -16,7 +16,7 @@ function handler(event, context, callback) {
     .then((token) => {
       console.log(`Logged in successfully, token: ${token}`);
       query(token, `
-        query { 
+        query {
           repository(owner: "${owner}", name: "${name}") {
             name
             issues(states: OPEN) { totalCount }
