+++ oscarbot/intents/describeLastCommit.js
@@ -1,8 +1,8 @@
 const moment = require('moment');
-const config = require('./config');
-const dialogActions = require('./utils/dialogActions');
-const github = require('./utils/github');
-const i18n = require('./i18n');
+const config = require('../config');
+const dialogActions = require('../utils/dialogActions');
+const github = require('../utils/github');
+const i18n = require('../i18n');
 
 function handler(event, context, callback) {
   const repository = event.sessionAttributes.Repository;
