+++ oscarbot/intents/index.js
@@ -1,0 +1,14 @@
+module.exports = {
+  countIssues: require('./countIssues'),
+  countOpenPRs: require('./countOpenPRs'),
+  describeLastCommit: require('./describeLastCommit'),
+  getStars: require('./getStars'),
+  openIssue: require('./openIssue'),
+  topIssues: require('./topIssues'),
+  starProject: require('./starProject'),
+  unstarProject: require('./unstarProject'),
+  whatAreMyMostPopularProjects: require('./whatAreMyMostPopularProjects'),
+  whatCanIAskYou: require('./whatCanIAskYou'),
+  whatProjectsAmIWorkingOn: require('./whatProjectsAmIWorkingOn'),
+  myOpenIssues: require('./myOpenIssues'),
+};
