+++ oscarbot/intents/starProject.js
@@ -1,8 +1,8 @@
-const dialogActions = require('./utils/dialogActions');
-const elicitSlot = require('./utils/dialog/elicitSlot');
-const failed = require('./utils/dialog/failed');
-const github = require('./utils/github');
-const i18n = require('./i18n');
+const dialogActions = require('../utils/dialogActions');
+const elicitSlot = require('../utils/dialog/elicitSlot');
+const failed = require('../utils/dialog/failed');
+const github = require('../utils/github');
+const i18n = require('../i18n');
 
 function checkConfirmationStatus(event, callback) {
   const confirmationStatus = event.currentIntent.confirmationStatus;
