+++ oscarbot/intents/tests/get-stars-with-slots.json
