+++ oscarbot/intents/tests/get-stars.json
