+++ oscarbot/intents/tests/open-issue-with-slots.json
